chore(agents): restructure agent_platform into canonical product layout
Move the flat backend into the canonical product structure from
products/architecture.md — presentation/ (DRF views + serializers),
logic/ (janitor client + spec schema), and tests/ — add the
per-package turbo.json, and set the owning team to team-agents.

Pure structural change: no runtime behavior changes, and generated
OpenAPI/MCP output is byte-identical (drf-spectacular keys off class
and basenames, not module paths). The facade/contracts layer is
intentionally deferred — agent_platform has no in-process Python
callers, so the OpenAPI surface already is its contract.
