+++ products/agent_platform/backend/tests/test_cron_fire.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 from posthog.test.base import APIBaseTest
 from unittest.mock import MagicMock, patch
 
-from .models import AgentApplication, AgentRevision
+from ..models import AgentApplication, AgentRevision
 
 
 class TestCronFireAction(APIBaseTest):
@@ -56,7 +56,7 @@ class TestCronFireAction(APIBaseTest):
             f"/revisions/{self.revision.id}/cron/fire/"
         )
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_cron_fire_passes_cron_name_to_janitor_and_returns_response(self, mock_janitor: MagicMock) -> None:
         mock_janitor.return_value.cron_fire.return_value = {
             "ok": True,
@@ -73,7 +73,7 @@ class TestCronFireAction(APIBaseTest):
             str(self.revision.id), cron_name="digest", request_id="click-1"
         )
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_cron_fire_request_id_is_optional(self, mock_janitor: MagicMock) -> None:
         # Without request_id the janitor mints a UUID; verify we forward None
         # so the janitor falls through to its own generator rather than us
@@ -91,13 +91,13 @@ class TestCronFireAction(APIBaseTest):
             str(self.revision.id), cron_name="digest", request_id=None
         )
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_cron_fire_rejects_missing_cron_name(self, mock_janitor: MagicMock) -> None:
         res = self.client.post(self.url, {}, format="json")
         self.assertEqual(res.status_code, 400)
         mock_janitor.return_value.cron_fire.assert_not_called()
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_cron_fire_rejects_empty_cron_name(self, mock_janitor: MagicMock) -> None:
         res = self.client.post(self.url, {"cron_name": ""}, format="json")
         self.assertEqual(res.status_code, 400)
