+++ products/agent_platform/backend/tests/test_fleet_stats.py
@@ -14,8 +14,8 @@ from unittest.mock import patch
 
 from rest_framework import status
 
-from .api import AgentApplicationViewSet, AgentFleetViewSet
-from .models import AgentApplication
+from ..models import AgentApplication
+from ..presentation.views import AgentApplicationViewSet, AgentFleetViewSet
 
 
 class TestAgentApplicationStats(APIBaseTest):
@@ -37,7 +37,7 @@ class TestAgentApplicationStats(APIBaseTest):
         )
         self.url = f"/api/projects/{self.team.id}/agent_applications/{self.application.id}/stats/"
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_stats_forwards_to_janitor_with_since(self, mock_janitor) -> None:
         mock_janitor.return_value.aggregate_for_application.return_value = {
             "liveCount": 2,
@@ -54,7 +54,7 @@ class TestAgentApplicationStats(APIBaseTest):
             since="2026-05-28T00:00:00Z",
         )
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_stats_omits_since_when_not_supplied(self, mock_janitor) -> None:
         # The janitor side picks the default `since` window (24h) when the
         # query param is absent — Django shouldn't fabricate one and shouldn't
@@ -88,7 +88,7 @@ class TestAgentFleetViewSet(APIBaseTest):
         self.url_stats = f"/api/projects/{self.team.id}/agent_fleet/stats/"
         self.url_live = f"/api/projects/{self.team.id}/agent_fleet/live_sessions/"
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_stats_forwards_team_id_and_since(self, mock_janitor) -> None:
         mock_janitor.return_value.aggregate_for_team.return_value = {
             "liveCount": 1,
@@ -104,7 +104,7 @@ class TestAgentFleetViewSet(APIBaseTest):
             since="2026-05-28T00:00:00Z",
         )
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_live_sessions_forwards_team_id_and_limit(self, mock_janitor) -> None:
         mock_janitor.return_value.list_live_for_team.return_value = {"results": []}
         resp = self.client.get(self.url_live, {"limit": 25})
@@ -114,7 +114,7 @@ class TestAgentFleetViewSet(APIBaseTest):
             limit=25,
         )
 
-    @patch("products.agent_platform.backend.api._janitor")
+    @patch("products.agent_platform.backend.presentation.views._janitor")
     def test_live_sessions_without_limit_defaults_to_none(self, mock_janitor) -> None:
         # The janitor picks its own 100-default; Django shouldn't impose a
         # different one.
