+++ products/agent_platform/backend/tests/test_revisions_slug_lookup.py
@@ -19,7 +19,7 @@ from posthog.test.base import APIBaseTest
 
 from rest_framework import status
 
-from .models import AgentApplication, AgentRevision
+from ..models import AgentApplication, AgentRevision
 
 
 class TestRevisionsSlugLookup(APIBaseTest):
