+++ products/agent_platform/backend/tests/test_slug_validation.py
@@ -13,7 +13,7 @@ from posthog.test.base import APIBaseTest
 from parameterized import parameterized
 from rest_framework import status
 
-from .models import AgentApplication, AgentRevision
+from ..models import AgentApplication, AgentRevision
 
 _DBS = {
     "default",
