+++ products/agent_platform/product.yaml
@@ -1,3 +1,3 @@
 name: Agent stack
 owners:
-    - team-devex
+    - team-agents
