+++ products/agent_platform/turbo.json
@@ -1,0 +1,3 @@
+{
+    "extends": ["//"]
+}
